src/ticketless_parking_system/edge/requirements.txt
opencv-python
numpy
nats-py
ultralytics
easyocr